requirements.txt
crewai
langchain-groq
crewai-tools
streamlit
python-dotenv
pandas
beautifulsoup4 
playwright
flask